move to parents items common to all toolchains
